requirements.txt
smolagents>=0.1.0
python-dotenv>=1.0.0
jupyter>=1.0.0
transformers>=4.37.0
torch>=2.2.0
accelerate>=0.27.0 